[PoC] 406056 build against maven 3.1-SNAPSHOT
- upgrade all library versions coming from maven core
  to versions used by maven 3.1-SNAPSHOT
- Remove direct dependencies to aether and sisu as far as possible
  (replace with maven core/plexus APIs)
- I can successfully build tycho-demo itp04 example
  project using both maven 3.1.0-alpha-1 and 3.0.4 using this tycho patch

TODOs:
 - analyse and fix currently @Ignored failing tests
 - check if P2RepositoryConnectorFactory still works with
   old maven version (it seems so but not sure/what is this needed for
   anyway?)

Bug: 406056

+++ tycho-core/src/main/java/org/eclipse/tycho/core/p2/P2RepositoryConnectorFactory.java
@@ -11,17 +11,17 @@
 package org.eclipse.tycho.core.p2;
 
 import org.codehaus.plexus.component.annotations.Component;
-import org.sonatype.aether.RepositorySystemSession;
-import org.sonatype.aether.repository.RemoteRepository;
-import org.sonatype.aether.spi.connector.RepositoryConnector;
-import org.sonatype.aether.spi.connector.RepositoryConnectorFactory;
-import org.sonatype.aether.transfer.NoRepositoryConnectorException;
+import org.eclipse.aether.RepositorySystemSession;
+import org.eclipse.aether.repository.RemoteRepository;
+import org.eclipse.aether.spi.connector.RepositoryConnector;
+import org.eclipse.aether.spi.connector.RepositoryConnectorFactory;
+import org.eclipse.aether.transfer.NoRepositoryConnectorException;
 
 @Component(role = RepositoryConnectorFactory.class, hint = "p2")
 public class P2RepositoryConnectorFactory implements RepositoryConnectorFactory {
 
-    public int getPriority() {
-        return 0;
+    public float getPriority() {
+        return 0f;
     }
 
     public RepositoryConnector newInstance(RepositorySystemSession session, RemoteRepository repository)
